feat: 선택된 랜덤 로또 화면에 출력
