+++ index.html
@@ -22,6 +22,8 @@
           </div>
           <div class="puchase-error-message"></div>
         </form>
+        <div class="random-lottos">
+        </div>
       </div>
       
     </main>
